requirements.txt
PyOpenGL~=3.1.7
pygame~=2.5.2